tidy_up: update documentation of the app.
update the documentation and comments to be more accurate and consistent,
as well as to better follow effective dart guidelines.

see https://dart.dev/guides/language/effective-dart/documentation.

this commit includes a couple fixups:
removing logic from state constructors
as a style ignore was uncommented,
and renaming a misspelt folder.